+++ library/src/test/java/com/cookieinformation/mobileconsents/sdk/MobileConsentSdkTest.kt
@@ -1,6 +1,7 @@
 package com.cookieinformation.mobileconsents.sdk
 
 import com.cookieinformation.mobileconsents.CallListener
+import com.cookieinformation.mobileconsents.CallbackMobileConsentSdk
 import com.cookieinformation.mobileconsents.Consent
 import com.cookieinformation.mobileconsents.ConsentSolution
 import com.cookieinformation.mobileconsents.MobileConsentSdk
@@ -51,7 +52,7 @@ internal class MobileConsentSdkTest : DescribeSpec({
   lateinit var baseUrl: HttpUrl
   lateinit var consentClient: ConsentClient
   lateinit var storage: ConsentStorage
-  lateinit var consentSdk: MobileConsentSdk
+  lateinit var consentSdk: CallbackMobileConsentSdk
   lateinit var consentSolution: ConsentSolution
 
   beforeTest {
@@ -62,10 +63,13 @@ internal class MobileConsentSdkTest : DescribeSpec({
     consentClient = ConsentClient(baseUrl, baseUrl, OkHttpClient(), moshi)
     storage = ConsentStorage(Mutex(), tempfile(suffix = ".txt"), MoshiFileHandler(moshi))
     consentSolution = ConsentSolutionResponseJsonAdapter(moshi).fromJson(consentString)!!.toDomain()
-    consentSdk = MobileConsentSdk(
-      consentClient = consentClient,
-      consentStorage = storage,
-      applicationProperties = applicationProperties,
+    consentSdk = CallbackMobileConsentSdk(
+      MobileConsentSdk(
+        consentClient = consentClient,
+        consentStorage = storage,
+        applicationProperties = applicationProperties,
+        dispatcher = Dispatchers.Unconfined
+      ),
       dispatcher = Dispatchers.Unconfined
     )
   }
@@ -151,7 +155,7 @@ private const val malformedJson =
 
 internal fun Class<*>.getResourceAsString(path: String) = getResourceAsStream(path)!!.readBytes().decodeToString()
 
-internal suspend fun MobileConsentSdk.getConsentSuspending(consentId: UUID): ConsentSolution =
+internal suspend fun CallbackMobileConsentSdk.getConsentSuspending(consentId: UUID): ConsentSolution =
   suspendCoroutine { continuation ->
     fetchConsentSolution(
       consentSolutionId = consentId,
@@ -167,7 +171,7 @@ internal suspend fun MobileConsentSdk.getConsentSuspending(consentId: UUID): Con
     )
   }
 
-internal suspend fun MobileConsentSdk.postConsentSuspending(consent: Consent) =
+internal suspend fun CallbackMobileConsentSdk.postConsentSuspending(consent: Consent) =
   suspendCoroutine<Unit> { continuation ->
     postConsent(
       consent = consent,
@@ -183,7 +187,7 @@ internal suspend fun MobileConsentSdk.postConsentSuspending(consent: Consent) =
     )
   }
 
-internal suspend fun MobileConsentSdk.getConsentChoicesSuspending() =
+internal suspend fun CallbackMobileConsentSdk.getConsentChoicesSuspending() =
   suspendCoroutine<Map<UUID, Boolean>> { continuation ->
     getSavedConsents(
       listener = object : CallListener<Map<UUID, Boolean>> {
@@ -198,7 +202,7 @@ internal suspend fun MobileConsentSdk.getConsentChoicesSuspending() =
     )
   }
 
-internal suspend fun MobileConsentSdk.getConsentChoiceSuspending(consentId: UUID) =
+internal suspend fun CallbackMobileConsentSdk.getConsentChoiceSuspending(consentId: UUID) =
   suspendCoroutine<Boolean> { continuation ->
     getSavedConsent(
       consentItemId = consentId,